Switch to Meson for the build system
This commit adds a meson build system, and removes the old CMake one. Griddyn
had ~10k lines of CMake over 101 files, for compilation rules of ~450
cpp files. A port of most of the compile rules to Meson yields ~800 lines, which
is much closer to the ideal 1 line per file.

This commit also builds all dependencies in ThirdParty manually, so that in-tree
changes are picked up without a full recompile, and so that things like
AddressSanitizer are usable on more of the codebase.

Finally, this commit splits up the tests into individual executables. There are
now 58 tests which can fail independently. The current test matrix looks as
follows:

Ok:                   17
Expected Fail:         0
Fail:                 41
Unexpected Pass:       0
Skipped:               0
Timeout:               0

As a result, this can't be a replacement for the CMake system, not yet.

This commit breaks all functionality which depends on co-simulation-related
libraries (fskit, fmi, fncs, helics). I can't test them, not having the required
dependencies.

Miscellanea:
 * Reorganize subfolders of ThirdParty for include isolation
 * Add cppzmq, yaml-cpp (minus tests) to follow other third party libraries
 * Move headers used between sub-libraries to include/, for isolation
 * Correct many include statements, to their new spots
 * Transitive dependencies are not represented
 * Boost dependencies are not split by actual requirement
 * Deleted in-tree tests