Fix leak in kinsolInterface
KINInit allocates some memory each time it runs, but does not free memory it
previously allocated. GridDyn reinitializes the solver a few times in the error
recovery bits of the codebase, so this was causing leaks. Fix that by not
reinitializing the solver.

Also, disable the n-2 contingency test. It's too long.